+++ NEWS.md
@@ -6,6 +6,8 @@
 - Add msLevel argument to chromatogram,MSnExp method <2017-08-16 Wed>
 - `calculateFragments` now just calculate fragments for all `n - 1L` bonds
     (before it incorrectly adds an additional bond; fixes #248) <2017-08-20 Sun>
+- Add `isEmpty` methods for `Chromatogram` and `Chromatograms` objects
+  <2017-09-05 Tue>
 
 ## Changes in version 2.3.9
 - Using new mzR::openIdfile backend to add identifcation data to raw